Start to change the output to be an html page.
Previously this project was using mermaid CLI to generate diagrams (both
pngs and svgs). This worked well, but was a little hard to consume. It
also didn't have a great interface to see what story was what. You would
have to visit links (which only worked in svgs). Although there were
tool tips setup, that only works in Mermaid in the browser (not in
svgs).
