+++ .env.sample
@@ -1,5 +1,8 @@
 # Find instructions here: https://www.pivotaltracker.com/help/articles/api_token/
 PIVOTAL_TRACKER_TOKEN="<>"
 
+# The project ID can be found in the url or the settings page
 PROJECT_ID=""
+
+# This is the text of the label
 EPIC_LABEL=""
