+++ .gitignore
@@ -4,3 +4,4 @@ node_modules
 *.mmd
 *.png
 *.svg
+epic_diagram.html
