+++ .nvmrc
@@ -1,1 +1,1 @@
-12.16.2
+16.13.2
